TGB-5 исправления 2
Привести названия к общему виду
Поместить папки в scr
Добавить .env.example И скрыть настоящий
Создать docs и убрать картинку туда
заменить на copy . .
Сделать key в БД
Переместить код из слушателя в папку/метод Sql
Убрать внутренний id
Перенести планировщик
