+++ .env
@@ -1,0 +1,6 @@
+TOKEN=7772980080:AAGwUXoFhXXAaIJZtzp2XGSNm2B3hoUIXZU
+DB_HOST=postgres
+DB_NAME=postgres
+DB_USER=postgres
+DB_PASSWORD=postgres
+Test_chat=719467479
