+++ .env.example
@@ -1,4 +1,6 @@
-TOKEN=7772980080:AAGwUXoFhXXAaIJZtzp2XGSNm2B3hoUIXZU
+#Токен бота
+TOKEN=token
+
 DB_HOST=postgres
 DB_NAME=postgres
 DB_USER=postgres
