+++ README.md
@@ -1,2 +1,2 @@
 # tg-bot-listener
-![alt text](image.png)
+![alt text](.\docs\image.png)
