+++ docker-compose.yml
@@ -11,6 +11,4 @@ services:
     depends_on:
       - postgres
     build: .
-    env_file: ./environment_variables.env
-    ports:
-      - "5000:5000"
+    env_file: ./.env
